+++ gemrb/plugins/ACMReader/unpacker.cpp
@@ -239,7 +239,7 @@ int CValueUnpacker::t1_5bits(int pass, int /*ind*/)
 	// efficiency: always 5/3 bits per value
 	// use it when P0 <= 1/3
 	for (int i = 0; i < subblocks; i++) {
-		int bits = ( int ) ( get_bits( 5 ) & 0x1f );
+		int bits = get_bits(5) & 0x1f;
 		bits = ( int ) Table1[bits];
 
 		block_ptr[i * sb_size + pass] = buff_middle[-1 + ( bits & 3 )];
@@ -312,7 +312,7 @@ int CValueUnpacker::t2_7bits(int pass, int /*ind*/)
 	// эффективность: 7/3 бита на значение - всегда
 	// use it when p0 <= 1/3
 	for (int i = 0; i < subblocks; i++) {
-		int bits = ( int ) ( get_bits( 7 ) & 0x7f );
+		int bits = get_bits(7) & 0x7f;
 		short val = Table2[bits];
 
 		block_ptr[i * sb_size + pass] = buff_middle[-2 + ( val & 7 )];
@@ -445,7 +445,7 @@ int CValueUnpacker::t3_7bits(int pass, int /*ind*/)
 	//Rus: все комбинации пар от -5 до +5
 	// эффективность: 7/2 бита на значение - всегда
 	for (int i = 0; i < subblocks; i++) {
-		int bits = ( int ) ( get_bits( 7 ) & 0x7f );
+		int bits = get_bits(7) & 0x7f;
 		unsigned char val = Table3[bits];
 
 		block_ptr[i * sb_size + pass] = buff_middle[-5 + ( val & 0xF )];
